+++ packages/widget/src/domain/types/connectors.ts
@@ -1,3 +1,6 @@
+import { isExternalProviderConnector } from "@sk-widget/providers/external-provider";
+import { isLedgerLiveConnector } from "@sk-widget/providers/ledger/ledger-live-connector-meta";
+import { isSafeConnector } from "@sk-widget/providers/safe/safe-connector-meta";
 import type { Observable } from "rxjs";
 import type { Chain } from "viem";
 import type { Connector } from "wagmi";
@@ -12,3 +15,11 @@ export const isConnectorWithFilteredChains = (
   return !!(connector as unknown as ConnectorWithFilteredChains)
     .$filteredChains;
 };
+
+export const shouldShowDisconnect = (connector: Connector) =>
+  !isExternalProviderConnector(connector) &&
+  !isLedgerLiveConnector(connector) &&
+  !isSafeConnector(connector);
+
+export const shouldMultiSend = (connector: Connector) =>
+  isSafeConnector(connector);