+++ packages/widget/src/pages/steps/hooks/use-steps-machine.hook.ts
@@ -1,3 +1,4 @@
+import { shouldMultiSend } from "@sk-widget/domain/types/connectors";
 import { useSavedRef } from "@sk-widget/hooks";
 import type { ActionDto, TransactionDto } from "@stakekit/api-hooks";
 import {
@@ -21,7 +22,6 @@ import {
   transactionsForConstructOnlySet,
 } from "../../../domain";
 import { useTrackEvent } from "../../../hooks/tracking/use-track-event";
-import { isExternalProviderConnector } from "../../../providers/external-provider";
 import { useSKWallet } from "../../../providers/sk-wallet";
 import type { GetStakeSessionError, SendTransactionError } from "./errors";
 import {
@@ -85,20 +85,15 @@ export const useStepsMachine = ({
   const transactionGetGasForNetwork = useTransactionGetGasForNetworkHook();
   const transactionConstruct = useTransactionConstructHook();
 
-  const shouldMultiSend = useMemo(
-    () =>
-      !!(
-        connector &&
-        isExternalProviderConnector(connector) &&
-        connector.shouldMultiSend
-      ),
+  const multiSend = useMemo(
+    () => Maybe.fromNullable(connector).map(shouldMultiSend).orDefault(false),
     [connector]
   );
 
   const machineParams = useSavedRef({
     transactions,
     integrationId,
-    shouldMultiSend,
+    multiSend,
     isLedgerLive,
     trackEvent,
     signMultipleTransactions,
@@ -120,7 +115,7 @@ const getMachine = (
     MutableRefObject<{
       transactions: ActionDto["transactions"];
       integrationId: ActionDto["integrationId"];
-      shouldMultiSend: boolean;
+      multiSend: boolean;
       isLedgerLive: boolean;
       trackEvent: ReturnType<typeof useTrackEvent>;
       signMultipleTransactions: ReturnType<
@@ -155,7 +150,7 @@ const getMachine = (
   const initContext = getInitContext(
     ref.current.transactions,
     ref.current.integrationId,
-    ref.current.shouldMultiSend
+    ref.current.multiSend
   );
 
   return setup({
@@ -274,7 +269,7 @@ const getMachine = (
               /**
                * Multi sign transactions
                */
-              if (ref.current.shouldMultiSend) {
+              if (ref.current.multiSend) {
                 return EitherAsync.liftEither(
                   Right(
                     txs.find((tx) =>