+++ packages/widget/vite/vite.config.base.ts
@@ -3,10 +3,12 @@ import { vanillaExtractPlugin } from "@vanilla-extract/vite-plugin";
 import react from "@vitejs/plugin-react-swc";
 import autoprefixer from "autoprefixer";
 import merge from "lodash.merge";
+import assignLayer from "postcss-assign-layer";
 import macros from "unplugin-parcel-macros";
 import type { UserConfig } from "vite";
 import { nodePolyfills } from "vite-plugin-node-polyfills";
-import type { InlineConfig } from "vitest";
+import type { InlineConfig } from "vitest/node";
+import { globalSKLayer } from "../src/styles/theme/ids";
 
 declare module "vite" {
   interface UserConfig {
@@ -36,11 +38,26 @@ export const getConfig = (overides?: Partial<UserConfig>): UserConfig =>
     ],
     css: {
       postcss: {
-        plugins: [autoprefixer()],
+        plugins: [
+          autoprefixer(),
+          assignLayer([
+            {
+              include: "**/*.css",
+              layerName: globalSKLayer,
+            },
+          ]),
+        ],
       },
     },
     esbuild: { drop: ["console"] },
-    server: { host: true },
+    server: {
+      host: true,
+      cors: true,
+      // https: {
+      //   key: path.resolve(__dirname, "..", "certificates", "skwidget.key"),
+      //   cert: path.resolve(__dirname, "..", "certificates", "skwidget.crt"),
+      // },
+    },
     resolve: {
       alias: {
         "@sk-widget": path.resolve(__dirname, "..", "src"),